+++ app/code/ThirdParty/BlogArticle/Model/Source/IsActive.php
@@ -1,0 +1,15 @@
+<?php
+namespace ThirdParty\BlogArticle\Model\Source;
+
+use Magento\Framework\Data\OptionSourceInterface;
+
+class IsActive implements OptionSourceInterface
+{
+    public function toOptionArray()
+    {
+        return [
+            ['value' => 1, 'label' => __('Enabled')],
+            ['value' => 0, 'label' => __('Disabled')],
+        ];
+    }
+}