+++ app/code/ThirdParty/BlogArticle/Ui/Component/Listing/Column/PostActions.php
@@ -1,0 +1,52 @@
+<?php
+namespace ThirdParty\BlogArticle\Ui\Component\Listing\Column;
+
+use Magento\Framework\UrlInterface;
+use Magento\Framework\View\Element\UiComponent\ContextInterface;
+use Magento\Framework\View\Element\UiComponentFactory;
+use Magento\Ui\Component\Listing\Columns\Column;
+
+class PostActions extends Column
+{
+    public const URL_PATH_EDIT = 'blogarticle/post/edit';
+    public const URL_PATH_DELETE = 'blogarticle/post/delete';
+
+    private $urlBuilder;
+
+    public function __construct(
+        ContextInterface $context,
+        UiComponentFactory $uiComponentFactory,
+        UrlInterface $urlBuilder,
+        array $components = [],
+        array $data = []
+    ) {
+        $this->urlBuilder = $urlBuilder;
+        parent::__construct($context, $uiComponentFactory, $components, $data);
+    }
+
+    public function prepareDataSource(array $dataSource)
+    {
+        if (!isset($dataSource['data']['items'])) {
+            return $dataSource;
+        }
+        foreach ($dataSource['data']['items'] as &$item) {
+            if (!isset($item['post_id'])) {
+                continue;
+            }
+            $name = $this->getData('name');
+            $item[$name]['edit'] = [
+                'href' => $this->urlBuilder->getUrl(self::URL_PATH_EDIT, ['post_id' => $item['post_id']]),
+                'label' => __('Edit'),
+            ];
+            $item[$name]['delete'] = [
+                'href' => $this->urlBuilder->getUrl(self::URL_PATH_DELETE, ['post_id' => $item['post_id']]),
+                'label' => __('Delete'),
+                'confirm' => [
+                    'title' => __('Delete post'),
+                    'message' => __('Are you sure you want to delete this post?'),
+                ],
+            ];
+        }
+        return $dataSource;
+    }
+}